pyproject.toml
[project]
name = "imagineui"
version = "0.1.0"
description = "An AI-powered CSS style generator using CSS Zen Garden designs"
authors = [
    {name = "Anthony Chapman", email = "adc@thewebdev.us"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "jupyter",
    "notebook",
    "ipykernel>=6.29.0",
    "beautifulsoup4>=4.12.0",
    "scrapy>=2.11.0",
    "playwright>=1.42.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.19",
    "chainlit>=0.7.700",
    "transformers>=4.37.0",
    "datasets>=2.17.0",
    "torch==2.1.2",
    "sentence-transformers>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.1",
    "openai>=1.63.2,<2.0.0",
    "langchain-openai>=0.3.6,<0.4.0",
    "langchain-community>=0.3.18,<0.4.0",
    "faiss-cpu>=1.7.4",
    "jq>=1.8.0",
    "ragas>=0.2.13",
    "rapidfuzz>=3.12.1",
    "websockets>=11.0.3",
    "langsmith>=0.3.11",
    "matplotlib>=3.10.0",
    "accelerate>=0.26.0",
    "nest-asyncio>=1.6.0",
    "anthropic>=0.49.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-ra -q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true 
